Upgrade libraries and add DesktopCompactScale for UI scaling
This commit updates the following libraries:
- composeMultiplatform to version 1.9.0
- m3-adaptive to version 1.2.0-alpha06

It also introduces `DesktopCompactScale`, a composable function that allows scaling of dp and sp units for desktop applications. This is used in the main window to scale the UI by 0.9f and disable minimum interactive component enforcement. The main window's minimum size is also set, and it is initially positioned in the center of the screen.